ci: add unit tests

+++ tests/AuthenticationEndpointTest.php
@@ -16,7 +16,9 @@ class AuthenticationEndpointTest extends TestCase
     public function test_process()
     {
         // clean old tests
-        unlink("tests/tests.sqlite");
+        if (file_exists("tests/tests.sqlite")) {
+            unlink("tests/tests.sqlite");
+        }
 
         $config = json_decode(file_get_contents(__DIR__ . '/../config_sample/config.json'), true);
         $database = new OP_Database(__DIR__ . '/tests.sqlite');
